// backend/src/modules/production-orders/dto/create-production-order-vial-inspection-check.dto.ts
export class CreateProductionOrderVialInspectionCheckDto {
  bag_number?: number | string | null;
  fiber_vial_count?: number | string | null;
  particulate_count?: number | string | null;
  damaged_count?: number | string | null;
  other_defect_count?: number | string | null;
  note?: string | null;
}